pyproject.toml
[project]
name = "qmost_mock_spectra"
dynamic = ["version"]
license = {text = "BSD 3-Clause"}
authors = [
  {name = "Harry Addison", email = "ha00871@surrey.ac.uk"}]
keywords = ["4MOST", "Spectra"]
description = "4MOST SN mock spectra"
readme = {file = "README.rst", content-type = "text/x-rst"}
dependencies = ["astropy", "extinction", "matplotlib", "numpy", "PyYAML"]

[project.optional-dependencies]
test = ["pytest", "pytest-doctestplus"]
doc = ["setuptools_scm", "sphinx", "sphinx-argparse", "sphinx-astropy"]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["qmost_mock_spectra*"]
namespaces = true

[tool.setuptools_scm]
write_to = "qmost_mock_spectra/version.py"